redesign problem search area
Current problem search area is highly unintuitive and lacks a couple of features discussed in last meeting.
To improve upon current design, make it more intuitive and pleasant looking, and to add new features, problem search area was redesigned and divided into three areas:
-Tag search: two collumns - one for searching for tags, and second for displaying and removing already selected tags.
-Difficulty select: dual slider allowing user to select range in which they want to search problems.
-Key words: input field allowing user to insert key words to search for problems.
The entire area was also adapted for use on small devices, as it was scaling badly before.
Some small changes to hintbase description to make it more readable were also added.

Tag search area and key words area have no backend. Difficulty slider has js backend that controls it's behaviour (not allowing min to be greater than max, coloring the range etc.), but it is not connected to database.