redesign seminars
The Seminar class was redefined to:
- separate audience information from difficulty,
- include special attributes for featured and special guest events,
- include useful information for API use like discord-channel.

New class SeminarGroup represents event audience and keeps default
difficulty for the audience to minimize changes in user experience.

Similar to Posts, moved formatting to the class.

Main page event info and /kolo calendar were modified to display the
added information to the user. In the process, polished the design and
fixed a number of known bugs.