build models for linked accounts and activity
This commit prepares the codebase for the implementation of bot API.

A new model was created for linked accounts, which aims to associate
external acounts from different platforms with a user. Make sure to
verify account ownership (e.g. via  OAuth) before adding any links.

Another model was created to store points that are soon to be given to
end users (particularly Discord members) for various activities. The
model allows to specify a reason why the user obtained the points. Each
user now has an activity score, which is a sum of the points given. To
avoid overloading the database and slow response times when Postgres is
queried for a user's score, further measures (such as periodically
aggregating old changes to reduce the number of rows per user) will have
to be implemented before full-scale deployment.