add admin page functionalities
The admin page was getting increasingly difficult to use, as there was
no way of searching or filtering the objects, which also were not sorted.

Seminars are now sorted by date decreasing and it is possible to filter
them by date range, as well as by group and tutors.

Seminar groups now display the number of seminars assigned to that group.

Staff users are now displayed before any other users and it is possible
to search for users by username, full name, or email.