Add debug toolbar to the project
Django debug toolbar (https://django-debug-toolbar.readthedocs.io/) is a
very useful tool that was already used multiple times by people contributing
to this repo, yet it was never added to the project.

Oficially added the debug toolbar as a dependecy. This change should
simplify development, eliminating the need for adding the tool locally every
time it is needed, and has no effect when debug=False.