Disable Redis by default when debugging
Redis was enabled by default even when debugging locally, yet the
instruction suggested to disable it, as it is not necessary unless
specific changes to caching are tested.

Modified the default behavior to disable Redis caching and simplify
local development. Redis can still be enabled in settings.py.

Updated the documentation to reflect this change.