Do not require db password for SQLite
Previously DB_PASSWORD had to be set even when using a local SQLite
database, which does not require a password. This might be confusing to
new contributors when setting up their environment.

Modified the imports to only import DB_PASSWORD when it is needed.

Updated the documentation to reflect this change.